LayoutTests/fast/css/ClientRect-attributes-prototype-expected.txt

Test that ClientRect's properties are on the prototype.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clientRect.__proto__ is ClientRect.prototype
PASS Object.getOwnPropertyNames(clientRect).length is 0
PASS Object.hasOwnProperty(clientRect, 'top') is false
PASS Object.hasOwnProperty(clientRect, 'right') is false
PASS Object.hasOwnProperty(clientRect, 'bottom') is false
PASS Object.hasOwnProperty(clientRect, 'left') is false
PASS Object.hasOwnProperty(clientRect, 'width') is false
PASS Object.hasOwnProperty(clientRect, 'height') is false

* clientRect.__proto__.top
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.top
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* clientRect.__proto__.right
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.right
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* clientRect.__proto__.bottom
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.bottom
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* clientRect.__proto__.left
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.left
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* clientRect.__proto__.width
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.width
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* clientRect.__proto__.height
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

* ClientRect.prototype.height
PASS descriptor.get is an instance of Function
PASS descriptor.enumerable is true
PASS descriptor.configurable is true

PASS successfullyParsed is true

TEST COMPLETE
